Use FLUSHDB instead of FLUSHALL.
This is a better way to delete only the databases which are used.
A even better way would be to use a prefix to delete only keys
created by Sengi. Maybe there are foreign data (not created
by Sengi) stored in the same Redis instance. So in this case
not the whole instance should be flushed.